Use the new location for template_renderer_factory
In pyramid 1.5 they moved this function out into another module.  This updates the code to try the new location if the old one fails.  In the long term, though, this dependency should be broken.  I just haven't quite figured out what the "correct" code is supposed to do or I'd submit that as a pull request instead.